Moved dashboard to root page
